+++ tests/test_cli.py
@@ -4,4 +4,4 @@ from data_project_manager import __version__
 
 
 def test_version():
-    assert __version__ == "1.1.1"
+    assert __version__ == "1.2.0"